Make plugin works in RedHat
The NvmePlugin class is responsible to get the info on /sys/block/ about the nvme devices whether the nvme-cli package is installed or not.
It was implemented a separate class for RedHat and Debian plataforms. For now, the only one tested was the RedHatNvmePlugin class.

The RedHatNvmePlugin class will "use" the add_copy_spec() method from NvmePlugin. It also check if the nvme-cli package is installed in the system, if so then it will pass some nvme commands to add_cmd_output() method.