Solve sosreport performance issue
The plugin was taking some seconds to run and the problem was related with two specific commands to get pci and slot location info using add_cmd_output(). To solve this, I used these info previously taken by the subprocess.check_output() command and put them as a file located in sos_strings/nvme/.

The get_spec() method was changed to copy_spec(), given that this name explains better what this method does.